MNT: move serverrpc to magic
